A mayo themed html format

A mayo themed html format

mayohtml(
  toc = FALSE,
  toc_float = TRUE,
  toc_depth = 6,
  number_sections = TRUE,
  ...
)

Arguments

toc
logical; TRUE to include a table of contents in the output

toc_float
TRUE to float the table of contents to the left of the main document content.

toc_depth
Depth of headers to include in table of contents

number_sections
logical; TRUE to number section headings

...
Additional arguments passed to rmarkdown::html_document